refactor: digging 페이지 컴포넌트 css 분리
