+++ src/app/digging/components/ArtistGrid/ArtistGrid.module.css
@@ -1,0 +1,51 @@
+.gridContainer {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(17rem, 1fr));
+  gap: 3rem 2.4rem;
+}
+
+.artistCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  cursor: pointer;
+  transition: transform 0.3s;
+  animation: slideUp 0.6s ease-out;
+}
+
+.artistCard:hover {
+  transform: translateY(-0.6rem);
+}
+
+.artistImgWrapper {
+  position: relative;
+  width: 15rem;
+  height: 15rem;
+  border-radius: 50%;
+  overflow: hidden;
+  margin-bottom: 1.4rem;
+  box-shadow: 0 1rem 2rem rgba(0, 0, 0, 0.3);
+  border: 2px solid rgba(255, 255, 255, 0.05);
+}
+
+.artistImg {
+  object-fit: cover;
+}
+
+.artistName {
+  font-size: 1.6rem;
+  font-weight: 600;
+  color: #fff;
+}
+
+@keyframes slideUp {
+  from {
+    opacity: 0;
+    transform: translateY(2rem);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
