+++ src/app/digging/components/ArtistGrid/ArtistGrid.tsx
@@ -1,18 +1,16 @@
 "use client";
 import Image from "next/image";
-import styles from "../Digging.module.css";
+import styles from "./ArtistGrid.module.css";
 
-export default function ArtistGrid({
-  artists,
-  onClick,
-}: {
-  artists: any[];
-  onClick: any;
-}) {
+export default function ArtistGrid({ artists, onClick }: { artists: any[]; onClick: any }) {
   return (
     <div className={styles.gridContainer}>
       {artists.map((item) => (
-        <div key={item.id} className={styles.artistCard} onClick={onClick}>
+        <div
+          key={item.id}
+          className={styles.artistCard}
+          onClick={onClick}
+        >
           <div className={styles.artistImgWrapper}>
             <Image
               src={item.image}
