+++ src/app/digging/components/PlaylistList/PlaylistList.module.css
@@ -1,0 +1,82 @@
+.listContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+
+.trackRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.2rem 2rem;
+  border-radius: 1.2rem;
+  transition: background 0.2s;
+  border: 1px solid transparent;
+}
+
+.trackRow:hover {
+  background: rgba(255, 255, 255, 0.08);
+  border-color: rgba(255, 255, 255, 0.05);
+}
+
+.trackLeading {
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+  flex: 1;
+  overflow: hidden;
+}
+
+.trackImgWrapper {
+  position: relative;
+  width: 5.2rem;
+  height: 5.2rem;
+  border-radius: 0.6rem;
+  overflow: hidden;
+  flex-shrink: 0;
+}
+
+.trackImg {
+  object-fit: cover;
+}
+
+.trackMeta {
+  display: flex;
+  flex-direction: column;
+  gap: 0.3rem;
+  overflow: hidden;
+}
+
+.trackMeta .name {
+  font-size: 1.6rem;
+  font-weight: 500;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.metaRowList {
+  display: flex;
+  align-items: center;
+  gap: 0.6rem;
+  font-size: 1.3rem;
+  color: rgba(255, 255, 255, 0.5);
+  margin-top: 0.4rem;
+}
+
+.trackTrailing {
+  display: flex;
+  align-items: center;
+  gap: 2.4rem;
+}
+
+.metaBadge {
+  display: flex;
+  align-items: center;
+  gap: 0.6rem;
+  padding: 0.6rem 1.2rem;
+  background: rgba(255, 255, 255, 0.05);
+  border-radius: 100px;
+  font-size: 1.2rem;
+  color: rgba(255, 255, 255, 0.6);
+}
