+++ src/app/digging/components/TrackList/TrackList.module.css
@@ -1,0 +1,115 @@
+.listContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+  animation: slideUp 0.6s ease-out;
+}
+
+.trackRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.2rem 2rem;
+  border-radius: 1.2rem;
+  transition: background 0.2s;
+  border: 1px solid transparent;
+}
+
+.trackRow:hover {
+  background: rgba(255, 255, 255, 0.08);
+  border-color: rgba(255, 255, 255, 0.05);
+}
+
+.trackLeading {
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+  flex: 1;
+  overflow: hidden;
+}
+
+.trackImgWrapper {
+  position: relative;
+  width: 5.2rem;
+  height: 5.2rem;
+  border-radius: 0.6rem;
+  overflow: hidden;
+  flex-shrink: 0;
+}
+
+.trackImg {
+  object-fit: cover;
+}
+
+.trackMeta {
+  display: flex;
+  flex-direction: column;
+  gap: 0.3rem;
+  overflow: hidden;
+}
+
+.trackMeta .name {
+  font-size: 1.6rem;
+  font-weight: 500;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.trackMeta .artist {
+  font-size: 1.4rem;
+  color: rgba(255, 255, 255, 0.5);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.trackTrailing {
+  display: flex;
+  align-items: center;
+  gap: 2.4rem;
+}
+
+.duration {
+  font-size: 1.4rem;
+  color: rgba(255, 255, 255, 0.3);
+  font-variant-numeric: tabular-nums;
+  min-width: 4rem;
+  text-align: right;
+}
+
+.trackActions {
+  display: flex;
+  gap: 1rem;
+}
+
+.actionBtn {
+  width: 3.6rem;
+  height: 3.6rem;
+  border-radius: 50%;
+  border: none;
+  background: transparent;
+  color: rgba(255, 255, 255, 0.3);
+  font-size: 1.8rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition: all 0.2s;
+}
+
+.actionBtn:hover {
+  color: #fff;
+  transform: scale(1.1);
+}
+
+@keyframes slideUp {
+  from {
+    opacity: 0;
+    transform: translateY(2rem);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
